// src/interFace/type.ts
export type TBlog = {
  id: number;
  image: StaticImageData;
  date: string;
  month?: string;
  blogTitle: string;
  desc?: string;
  blogTag?: string;
  authorName?: string;
  desination?: string;
  wowDelay?: string;
  authorImg?: StaticImageData;
  comment?: number;
  videoId?:string;
}